Implement NSGAIIIElitePopulationSelectionStrategy
