+++ optuna/samplers/_nsgaiii/__init__.py
@@ -1,0 +1,3 @@
+from optuna.samplers._nsgaiii.sampler import NSGAIIISampler
+
+__all__ = ["NSGAIIISampler"]
